Move code to own file
